Add GD32F4 platform support
Add GD32F460RG target, GD32F4 build integration, startup/linker files, USB, UART, I2C, SPI, ADC, SDIO, timer, DMA, DShot, MCO, and config-flash hooks.

Fix a DShot bit-band telemetry decoder fault observed during manual single-motor tests. When a non-running ESC produced too few telemetry samples, decode_bb_bitband() computed count - MIN_VALID_BBSAMPLES before validating count, underflowing the end pointer and potentially scanning invalid bit-band memory. Return DSHOT_TELEMETRY_NOEDGE before pointer arithmetic when the capture is too short; apply the same guard to the non-bit-band decoder.

Validated with: GCC_REQUIRED_VERSION=13.2.1 automation/scripts/run-build-matrix.sh GD32F460RG GD32H757VI

+++ src/main/pg/serial_uart.c
@@ -108,7 +108,7 @@ void pgResetFn_serialUartConfig(serialUartConfig_t *config)
 
     for (unsigned i = 0; i < ARRAYLEN(uartDmaopt); i++) {
         const int resourceIndex = serialResourceIndex(uartDmaopt[i].identifier);
-        if (resourceIndex >= 0 && resourceIndex < UARTDEV_CONFIG_MAX) {  // hadle corrupted config gracefuly
+        if (resourceIndex >= 0 && resourceIndex < (int)UARTDEV_CONFIG_MAX) {  // hadle corrupted config gracefuly
             config[resourceIndex].txDmaopt = uartDmaopt[i].txDmaopt;
             config[resourceIndex].rxDmaopt = uartDmaopt[i].rxDmaopt;
         }
